+++ src/operations/getFileContents.ts
